+++ app/Http/Controllers/AdminController.php
@@ -1,7 +1,7 @@
 <?php
 
 
-class AdminController extends BaseController
+class AdminController extends Controller
 {
 
     public function __construct()
